First upload of tool files – v1.0 Beta
This is version 1.0 Beta of the tool. It has:

One main .go file (the main code of the tool)

A Makefile to help build and set up the tool

A setup.sh file to set some tool options


This version is in beta and still needs more work.
I am using TERMUX to write and build the tool, so there may be small problems on other systems.